requirements.txt
# ============================================================================
# E-commerce Automated Resolution System - Dependencies
# ============================================================================

# ----------------------------------------------------------------------------
# API and Web Server
# ----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# ----------------------------------------------------------------------------
# WhatsApp Integration
# ----------------------------------------------------------------------------
pywa>=3.0.0
pywa[fastapi]

# ----------------------------------------------------------------------------
# AI/LLM Services
# ----------------------------------------------------------------------------
groq>=0.4.0

# ----------------------------------------------------------------------------
# Model Context Protocol (MCP)
# ----------------------------------------------------------------------------
mcp>=0.9.0

# ----------------------------------------------------------------------------
# Database
# ----------------------------------------------------------------------------
pymongo>=4.5.0

# ----------------------------------------------------------------------------
# Machine Learning and Deep Learning
# ----------------------------------------------------------------------------
# Note: Install PyTorch with CUDA support using:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For CPU-only: pip install torch torchvision torchaudio

torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
transformers==4.57.1
scikit-learn==1.7.2

# ----------------------------------------------------------------------------
# Data Processing and Analysis
# ----------------------------------------------------------------------------
pandas==2.3.3
numpy==2.3.4

# ----------------------------------------------------------------------------
# Visualization and Plotting
# ----------------------------------------------------------------------------
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# ----------------------------------------------------------------------------
# Dashboard (Streamlit)
# ----------------------------------------------------------------------------
streamlit>=1.28.0

# ----------------------------------------------------------------------------
# Development and Testing
# ----------------------------------------------------------------------------
# Jupyter Notebook Support
ipykernel==7.1.0
ipython==9.6.0

# ----------------------------------------------------------------------------
# Notes:
# ----------------------------------------------------------------------------
# - All standard library imports (os, sys, json, logging, asyncio, etc.) 
#   are built-in and don't need to be installed
# - For CUDA support, ensure NVIDIA drivers and CUDA toolkit are installed
# - Use Python 3.8 or higher
# - For production deployment, consider pinning all versions

